pyproject.toml
[project]
name = "investbot"
version = "0.1.0"
description = "Инвестиционный бот на базе T-Bank API"
authors = [
    {name = "Alexander", email = "aleksanderfranorge@gmail.com"}
]
requires-python = ">=3.11,<3.15"

dependencies = [
    "aiohttp>=3.13.5,<4.0.0",
    "aiohttp-socks>=0.11.0,<0.12.0",
    "aiogram>=3.27.0,<4.0.0",
    "t-tech-investments @ git+https://opensource.tbank.ru/invest/invest-python.git"
]

[dependency-groups]

dev = [
    "pytest>=9.0.2,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "pre-commit>=4.0.0"
]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_configs = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-bot = "investbot.main:main"